Made a random generator fully in C. Need to test and benchmark timing against FORTRAN. Also, currently Rutherford has default material parameters and sampling range; need to pass them as variables.

+++ xcoll/scattering_routines/everest/jaw.h
@@ -1,34 +1,32 @@
 #include <math.h>
 
-float iterat(float a, float b, float dh, float s) {
+double iterat(double a, double b, double dh, double s) {
 
-	float ds = s;
+    double ds = s;
 
     while (1) {
         
-		ds = ds*0.5;
+        ds = ds*0.5;
 
         if (pow(s,3) < pow((a+b*s),2)) {
             s = s+ds;
-		} 
-		else {
+        } else {
             s = s-ds; 
-		}
+        }
 
         if (ds < dh) {
             break;
-		} 
-		else { 
+        } else { 
             continue;
-		}
-	}
+        }
+    }
 
     return s;
 
 }
   
 
-float soln3(float a, float b, float dh, float smax, float s) {
+double soln3(double a, double b, double dh, double smax, double s) {
 
 
     if (b == 0) {
@@ -36,69 +34,60 @@ float soln3(float a, float b, float dh, float smax, float s) {
 
         if (s > smax) {
             s = smax;
-		}
+        }
         return s;
-	}
+    }
 
     if (a == 0) {    
         if (b > 0) {
             s = pow(b,2);
-		}
-        else {
+        } else {
             s = 0;
-		}
+        }
         if (s > smax) {
             s = smax;
-		}
+        }
         return s;
-	}
+    }
         
     if (b > 0) {
 
         if (pow(smax,3) <= pow((a + b*smax),2)) {
             s = smax;
             return s;
-		}
-      
-        else {
+        } else {
             s = smax*0.5;
             s = iterat(a,b,dh,s);
-		}
-	}
-    
-    else {
-        float c = (-1*a)/b;
+        }
+    } else {
+        double c = (-1*a)/b;
         if (smax < c) {
             if ((pow(smax,3)) <= pow((a + b*smax),2)) {
                 s = smax;
                 return s;
-			}
-
-            else {
+            } else {
                 s = smax*0.5;
                 s = iterat(a,b,dh,s);
-			}
-		}
-    
-        else {
+            }
+        } else {
             s = c*0.5;
             s = iterat(a,b,dh,s);
-		}
-	}
+        }
+    }
    
     return s;
 
 }
 
 
-// float mcs(float s, float mc_radl, float mc_zlm1, float mc_p0, float mc_x, float mc_xp, float mc_z, float mc_zp, float mc_dpop) {
+// double mcs(double s, double mc_radl, double mc_zlm1, double mc_p0, double mc_x, double mc_xp, double mc_z, double mc_zp, double mc_dpop) {
 
-//     float theta    = 13.6e-3/(mc_p0*(1+mc_dpop)); // dpop   = (p - p0)/p0;
-//     float h   = 0.001;
-//     float dh  = 0.0001;
-//     float bn0 = 0.4330127019;
-//     float rlen0 = mc_zlm1/mc_radl;
-//     float rlen  = rlen0;
+//     double theta    = 13.6e-3/(mc_p0*(1+mc_dpop)); // dpop   = (p - p0)/p0;
+//     double h   = 0.001;
+//     double dh  = 0.0001;
+//     double bn0 = 0.4330127019;
+//     double rlen0 = mc_zlm1/mc_radl;
+//     double rlen  = rlen0;
 
 //     mc_x     = (mc_x/theta)/mc_radl;
 //     mc_xp    = mc_xp/theta;
@@ -108,17 +97,17 @@ float soln3(float a, float b, float dh, float smax, float s) {
 
 //     while (1) {
         
-//         float ae = bn0*mc_x;
-//         float be = bn0*mc_xp;
+//         double ae = bn0*mc_x;
+//         double be = bn0*mc_xp;
 
 //         // #######################################
-//         // ae = np.array(ae, dtype=np.float64)
-//         // be = np.array(be, dtype=np.float64)
-//         // dh = np.array(dh, dtype=np.float64)
-//         // rlen = np.array(rlen, dtype=np.float64)
-//         // s = np.array(s, dtype=np.float64)
+//         // ae = np.array(ae, dtype=np.double64)
+//         // be = np.array(be, dtype=np.double64)
+//         // dh = np.array(dh, dtype=np.double64)
+//         // rlen = np.array(rlen, dtype=np.double64)
+//         // s = np.array(s, dtype=np.double64)
 //         // #######################################
-//         float s = soln3(ae,be,dh,rlen,s);
+//         double s = soln3(ae,be,dh,rlen,s);
 
 //         if (s < h) {
 //             s = h;
